mesa: Adjust install for files provided by imx-gpu-viv
- Remove wayland-egl.pc
- Remove KHR headers

(cherry picked from commit 36b1840a853ed1b2d766b06b6644692e7f95d398)

+++ recipes-graphics/mesa/mesa_%.bbappend
@@ -34,10 +34,9 @@ do_install_append_imxgpu3d () {
     fi
 }
 
-do_install_append_imxgpu2d () {
-    rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
-}
-
-do_install_append_imxgpu2d () {
-    rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
+do_install_append_imxgpu () {
+    rm -f ${D}${includedir}/KHR
+    if [ "${BACKEND}" = "wayland" ]; then
+        rm -f ${D}${libdir}/pkgconfig/wayland-egl.pc
+    fi
 }
